Rename the default EventEngine headers
Small cleanup. This code hasn't been related to factories for a month or
two.
