+++ build_autogenerated.yaml
@@ -748,8 +748,8 @@ libs:
   - src/core/lib/debug/stats_data.h
   - src/core/lib/debug/trace.h
   - src/core/lib/event_engine/channel_args_endpoint_config.h
+  - src/core/lib/event_engine/default_event_engine.h
   - src/core/lib/event_engine/default_event_engine_factory.h
-  - src/core/lib/event_engine/event_engine_factory.h
   - src/core/lib/event_engine/executor/executor.h
   - src/core/lib/event_engine/executor/threaded_executor.h
   - src/core/lib/event_engine/handle_containers.h
@@ -1934,8 +1934,8 @@ libs:
   - src/core/lib/debug/stats_data.h
   - src/core/lib/debug/trace.h
   - src/core/lib/event_engine/channel_args_endpoint_config.h
+  - src/core/lib/event_engine/default_event_engine.h
   - src/core/lib/event_engine/default_event_engine_factory.h
-  - src/core/lib/event_engine/event_engine_factory.h
   - src/core/lib/event_engine/executor/executor.h
   - src/core/lib/event_engine/executor/threaded_executor.h
   - src/core/lib/event_engine/handle_containers.h
