+++ grpc.gemspec
@@ -960,10 +960,10 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/debug/trace.h )
   s.files += %w( src/core/lib/event_engine/channel_args_endpoint_config.cc )
   s.files += %w( src/core/lib/event_engine/channel_args_endpoint_config.h )
+  s.files += %w( src/core/lib/event_engine/default_event_engine.h )
   s.files += %w( src/core/lib/event_engine/default_event_engine_factory.cc )
   s.files += %w( src/core/lib/event_engine/default_event_engine_factory.h )
   s.files += %w( src/core/lib/event_engine/event_engine.cc )
-  s.files += %w( src/core/lib/event_engine/event_engine_factory.h )
   s.files += %w( src/core/lib/event_engine/executor/executor.h )
   s.files += %w( src/core/lib/event_engine/executor/threaded_executor.cc )
   s.files += %w( src/core/lib/event_engine/executor/threaded_executor.h )
