+++ src/core/lib/event_engine/event_engine.cc
@@ -21,8 +21,8 @@
 
 #include <grpc/event_engine/event_engine.h>
 
+#include "src/core/lib/event_engine/default_event_engine.h"
 #include "src/core/lib/event_engine/default_event_engine_factory.h"
-#include "src/core/lib/event_engine/event_engine_factory.h"
 
 namespace grpc_event_engine {
 namespace experimental {
