+++ tools/doxygen/Doxyfile.c++.internal
@@ -1943,10 +1943,10 @@ src/core/lib/debug/trace.cc \
 src/core/lib/debug/trace.h \
 src/core/lib/event_engine/channel_args_endpoint_config.cc \
 src/core/lib/event_engine/channel_args_endpoint_config.h \
+src/core/lib/event_engine/default_event_engine.h \
 src/core/lib/event_engine/default_event_engine_factory.cc \
 src/core/lib/event_engine/default_event_engine_factory.h \
 src/core/lib/event_engine/event_engine.cc \
-src/core/lib/event_engine/event_engine_factory.h \
 src/core/lib/event_engine/executor/executor.h \
 src/core/lib/event_engine/executor/threaded_executor.cc \
 src/core/lib/event_engine/executor/threaded_executor.h \
